🚧 WIP: only offer outlets, actions, destinations if they are available
This approach doen not quite work, because the @Presentation macro does not have any information about anything in the Feature. Maybe a Presentation protocol with associated type for the Feature is the way to go? I could add conditional extensions to the protocol if the feature’s Outlets, Actions, and Destinations are not Never…